Changes data types from unsigned long long to unsigned int
Reduces memory usage by switching from 64-bit to 32-bit unsigned integers throughout the graph and disjoint set implementations.

Updates all variable declarations, function parameters, return types, and printf format specifiers to use unsigned int instead of unsigned long long for better memory efficiency in typical use cases.

+++ disjointset.c
@@ -1,16 +1,16 @@
 #include<stdlib.h>
 
 struct disjointSet{
-    unsigned long long n;
-    unsigned long long* parent;
+    unsigned int n;
+    unsigned int* parent;
     unsigned char* rank;
 };
 
-struct disjointSet newDisjointSet(const unsigned long long n){
+struct disjointSet newDisjointSet(const unsigned int n){
     struct disjointSet ds;
     ds.n = n;
-    ds.parent = (unsigned long long*)malloc((n + 1) * sizeof(unsigned long long));
-    for(unsigned long long i = 1; i <= n; i++){
+    ds.parent = (unsigned int*)malloc((n + 1) * sizeof(unsigned int));
+    for(unsigned int i = 1; i <= n; i++){
         ds.parent[i] = i;
     }
     ds.rank = (unsigned char*)calloc(n + 1, sizeof(unsigned char));
@@ -22,16 +22,16 @@ void freeDisjointSet(struct disjointSet* const ds){
     free(ds->rank);
 }
 
-unsigned long long find(struct disjointSet* const ds, const unsigned long long u){
+unsigned int find(struct disjointSet* const ds, const unsigned int u){
     if(ds->parent[u] != u){
         ds->parent[u] = find(ds, ds->parent[u]);
     }
     return ds->parent[u];
 }
 
-void unionSet(struct disjointSet* const ds, const unsigned long long u, const unsigned long long v){
-    unsigned long long rootU = find(ds, u);
-    unsigned long long rootV = find(ds, v);
+void unionSet(struct disjointSet* const ds, const unsigned int u, const unsigned int v){
+    unsigned int rootU = find(ds, u);
+    unsigned int rootV = find(ds, v);
     if(rootU == rootV) return;
     if(ds->rank[rootU] < ds->rank[rootV]){
         ds->parent[rootU] = rootV;
